test(mip): Removed RNA conftest
- Moved Tb Mock to workflows
- Added mip_run tests
