+++ tests/cli/workflow/mip_dna/test_cli_mip_dna_run.py
@@ -1,23 +1,23 @@
-""" Test the CLI for run mip-rna """
+""" Test the CLI for run mip-dna """
 import logging
 
-from cg.cli.workflow.mip_rna.base import run
+from cg.cli.workflow.mip_dna.base import run
 from cg.apps.mip import MipAPI
 
 CASE_ID = "yellowhog"
-CONDA_ENV = "S_mip_rd-rna"
+CONDA_ENV = "S_mip_rd-dna"
 EMAIL = "james.holden@scilifelab.se"
 
 
 def test_cg_dry_run(cli_runner, tb_api, analysis_store_single_case, caplog):
-    """Test print the MIP command to console"""
+    """Test print the MIP run to console"""
     # GIVEN a cli function
 
     context = {}
     context["db"] = analysis_store_single_case
     context["tb"] = tb_api
-    context["rna_api"] = MipAPI("${HOME}/bin/mip", "analyse rd_rna", CONDA_ENV)
-    context["mip-rd-rna"] = {"mip_config": "config.yaml"}
+    context["dna_api"] = MipAPI("${HOME}/bin/mip", "analyse rd_dna", CONDA_ENV)
+    context["mip-rd-dna"] = {"mip_config": "config.yaml"}
 
     # WHEN we run a case in dry run mode
     caplog.set_level(logging.INFO)
@@ -26,13 +26,13 @@ def test_cg_dry_run(cli_runner, tb_api, analysis_store_single_case, caplog):
     # THEN the command should be printed
     with caplog.at_level(logging.INFO):
         assert (
-            "${HOME}/bin/mip analyse rd_rna yellowhog --config config.yaml "
-            "--email james.holden@scilifelab.se" in caplog.text
+            "${HOME}/bin/mip analyse rd_dna yellowhog --config config.yaml "
+            "--slurm_quality_of_service normal --email james.holden@scilifelab.se" in caplog.text
         )
 
 
 def test_run(cli_runner, tb_api, analysis_store_single_case, caplog, monkeypatch):
-    """Test run MIP RNA analysis"""
+    """Test run MIP DNA analysis"""
 
     def mip_run(_self, **kwargs):
         """monkeypatch function so we don't actually run MIP"""
@@ -41,11 +41,11 @@ def test_run(cli_runner, tb_api, analysis_store_single_case, caplog, monkeypatch
     # GIVEN a cli function
     context = {}
     monkeypatch.setattr(MipAPI, "run", mip_run)
-    mip_api = MipAPI("${HOME}/bin/mip", "analyse rd_rna", CONDA_ENV)
+    mip_api = MipAPI("${HOME}/bin/mip", "analyse rd_dna", CONDA_ENV)
     context["db"] = analysis_store_single_case
     context["tb"] = tb_api
-    context["rna_api"] = mip_api
-    context["mip-rd-rna"] = {"mip_config": "config.yaml"}
+    context["dna_api"] = mip_api
+    context["mip-rd-dna"] = {"mip_config": "config.yaml"}
 
     # WHEN we run a case
     caplog.set_level(logging.INFO)
@@ -53,7 +53,7 @@ def test_run(cli_runner, tb_api, analysis_store_single_case, caplog, monkeypatch
 
     # THEN we should get to the end of the function
     with caplog.at_level(logging.INFO):
-        assert "MIP rd-rna run started!" in caplog.text
+        assert "MIP rd-dna run started!" in caplog.text
 
     # WHEN we run a case in MIP dry mode
     caplog.set_level(logging.INFO)
@@ -61,4 +61,4 @@ def test_run(cli_runner, tb_api, analysis_store_single_case, caplog, monkeypatch
 
     # THEN we should get to the end of the function
     with caplog.at_level(logging.INFO):
-        assert "MIP rd-rna run started!" in caplog.text
+        assert "MIP rd-dna run started!" in caplog.text
